[TASK 3] The truffle config changes
I did not want to commit this file on the first place as it contains too much of my personal information that I do not want to share (like mneumonic and my infuraKey). But the end effect that I have done it properly can be already seen in the Rinkeby contract address added in the read me file.